requirements.txt
streamlit
speedtest-cli
pillow